# API-SPECIFICATION.yml
openapi: 3.0.3
info:
  title: SwiftEats Backend API
  version: 0.1.0
  description: |
    HTTP API for SwiftEats food delivery platform.
    Base path: /api/v1
  contact:
    name: SwiftEats
servers:
  - url: http://localhost:3000
    description: Local
security:
  - {}
...
paths:
...
  /api/v1/auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterDto'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  email:
                    type: string
                    format: email
        '400':
          description: Validation error
  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: Login with email and password
      description: Uses local auth guard. On success returns a JWT access token.
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginDto'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
        '401':
          description: Unauthorized
...
  /api/v1/auth/google/callback:
    get:
      tags: [Auth]
      summary: Google OAuth callback
      operationId: googleCallback
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
        '401':
          description: Unauthorized
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    UserRole:
      type: string
      enum: [customer, restaurant_owner, driver, admin]
    RegisterDto:
      type: object
      required: [email, password, role, fullName]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        role:
          $ref: '#/components/schemas/UserRole'
        fullName:
          type: string
        phone:
          type: string
        address:
          type: string
    LoginDto:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
...